+++ test/visual/index.ts
@@ -1,2 +1,1 @@
-import './freedraw';
 import './svg_import';